+++ packages/application-shell/src/components/project-switcher/project-switcher.spec.tsx
@@ -21,7 +21,6 @@ const render = () => {
     {
       disableRoutePermissionCheck: true,
       mocks: mockedRequest,
-      enableApolloMocks: true,
     }
   );
 };