+++ packages/application-shell/src/components/service-page-project-switcher/service-page-project-switcher.spec.js
@@ -50,7 +50,6 @@ describe('rendering', () => {
             },
           },
           mocks: mockedRequest,
-          enableApolloMocks: true,
         }
       );
       // eslint-disable-next-line testing-library/prefer-find-by