+++ packages/application-shell/src/hooks/use-applications-menu/use-applications-menu.spec.tsx
@@ -198,7 +198,6 @@ describe('for production usage', () => {
             },
           },
         ],
-        enableApolloMocks: true,
       });
       await waitForElementToBeRemoved(() => screen.queryByText('loading'));
       expect(screen.getByText('Orders')).toBeInTheDocument();
@@ -229,7 +228,6 @@ describe('for production usage', () => {
               error,
             },
           ],
-          enableApolloMocks: true,
         }
       );
       await screen.findByText('loading');